Fix #11: After saving config the old config values are shown
We fix the immediate problem by applying the PRG pattern here, instead
of trying to load `config.php` what doesn't work as of CMSimple_XH
1.7.0, if the configuration is already loaded.

However, it appears to be generally a bad idea to mess with the
handling of config and language. It would be much cleaner and more
future proof to either rely solely on CMSimple_XH's handling, or to
have own config files and variables.