Fix #12: Weak password hashes in cookie
We must not store weak password hashes (SHA-1 is definitely weak, and
SHA-256 is still doubtful for password hashing) in a cookie, and of
course we must not use the password hash of the admin password as salt.

Instead of rolling our own password hashing and verification, we use
`password_hash()` and `password_verify()` which have been designed for
this purpose. We add a the well known password_compat as fallback for
PHP versions before 5.5.0 and CMSimple_XH versions before 1.7.0. This
still requires PHP ≥ 5.3.7, though.